refactor(ir): ✓ move the analysis leaves below the IR, bank 92 → 83 (#4601 route 1)
With `loop-shape.ts` below the IR, the `remainder-fast-path` /
`static-numeric-range` pair that #4601's progress log recorded as NOT leaf-provable
becomes one: `static-numeric-range`'s third edge — the `loopBodyMutatesIndexOrArray`
call added by 8e77e6740 — now resolves below the boundary instead of into
`src/codegen/`. Both move. `analysis/i32-coerced-locals.ts` was already a
`ts-api`-only leaf and moves with them; it is what empties `ir/analysis/i32-slots.ts`.

Pure motion, ten consumer import lines rewritten (6 IR-side, 10 codegen-side).
`prove-emit-identity check` IDENTICAL across all 60 (file,target) pairs; ts7
typecheck clean; check:ir-fallbacks / check:ir-only / check:linear-ir /
check:ir-dialect all unchanged.

check:ir-layering 92 → 83 import lines, 19 → 15 files, banked via --update.
Four files clear their codegen dependency entirely: ir/analysis/i32-slots.ts,
ir/char-read-loop.ts, ir/fixed-literal-loop-proof.ts, ir/remainder-fast-path.ts.
The gate reports NO new file with codegen imports — which is the leaf-proof the
earlier attempt could not produce, and the reason that attempt was not banked.

The `integration.ts` bridge (41 of the remaining 83 lines) is untouched; #4601
stays `blocked` on #3520/#3521 for that scope.


✓
