+++ src/ir/analysis/remainder-fast-path.ts
